feat: rename tracking to analytics
