+++ api/src/analytics/analytics.service.ts
@@ -1,9 +1,9 @@
 import { Injectable } from "@nestjs/common";
-import { TrackingEvent } from "@/tracking/type";
+import { TrackingEvent } from "@/analytics/type";
 import { ConfigService } from "@nestjs/config";
 
 @Injectable()
-export class TrackingService {
+export class AnalyticsService {
   constructor(private configService: ConfigService) {}
 
   async trackEvent(eventData: TrackingEvent) {
@@ -13,15 +13,17 @@ export class TrackingService {
     // todo understand the params
     // and add the host / value and other missing fields from the front end
 
-    // api refrence : https://developer.matomo.org/api-reference/tracking-api
+    // api reference : https://developer.matomo.org/api-reference/tracking-api
+    // as per doc, all parameters values that are strings (such as 'url', 'action_name', etc.) must be URL encoded.
     const params = new URLSearchParams({
       idsite: siteId,
       rec: "1",
       apiv: "1",
-      e_c: eventData.category,
-      e_a: eventData.action,
-      e_n: eventData.name,
-      // ... autres paramètres nécessaires
+      // todo remove suffixe server side once implementation has been tested
+      e_c: `${eventData.category}-server-side`,
+      e_a: `${eventData.action}-server-side`,
+      e_n: `${eventData.name}-server-side`,
+      e_v: eventData.value,
     });
 
     const response = await fetch(matomoUrl, {
