+++ api/src/analytics/type.ts
@@ -2,4 +2,5 @@ export interface TrackingEvent {
   name: string;
   action: string;
   category: string;
+  value: string;
 }
