+++ api/src/app.module.ts
@@ -18,7 +18,7 @@ import { ProjetQualificationModule } from "@/projet-qualification/projet-qualifi
 import { BullBoardModule } from "@bull-board/nestjs";
 import basicAuth from "express-basic-auth";
 import { ExpressAdapter } from "@bull-board/express";
-import { TrackingModule } from "@/tracking/tracking.module";
+import { AnalyticsModule } from "@/analytics/analytics.module";
 
 @Module({
   imports: [
@@ -62,7 +62,7 @@ import { TrackingModule } from "@/tracking/tracking.module";
     LoggerModule,
     GeoModule,
     ProjetQualificationModule,
-    TrackingModule,
+    AnalyticsModule,
   ],
   providers: [AppService, ThrottlerGuardProvider, RequestLoggingInterceptor],
 })
